requirements.txt
pandas>=1.0.0
google-generativeai>=0.1.0
python-dotenv>=0.10.0
tqdm>=4.0.0
chardet>=3.0.0